+++ app/controllers/comments_controller.rb
@@ -1,41 +1,30 @@
 class CommentsController < ApplicationController
-  before_action :set_comment, only: [:edit, :update, :destroy]
+  before_action :set_comment, only: [:destroy]
   before_action :set_negotiation
 
-  def new
-    @comment = Comment.new
-  end
-
   def create
-    if Comment.create(comment_params)
-      redirect_to root_path
+    @comment = @negotiation.comments.new(comment_params)
+    # binding.pry
+    if @comment.save
+      redirect_to negotiation_path(@negotiation.id)
     else
       render :new
     end
   end
 
-  def edit
-  end
-
-  def update
-    if @comment.update(comment_params)
-      redirect_to root_path
-    else
-      render :edit
-    end
-  end
-
   def destroy
-    if @comment.destroy
-      redirect_to  root_path
-    else
-      render :edit
+    if @comment.user_id == current_user.id
+      if @comment.destroy
+        redirect_to negotiation_path(@negotiation.id)
+      else
+        render :edit
+      end
     end
   end
 
   private
   def comment_params
-    params.require(:comment).permit(:body).merge(user_id: current_user.id, negotiation_id: @negotiation.id)
+    params.permit(:body).merge(user_id: current_user.id)
   end
 
   def set_negotiation
